Move validation utilities to utils
